chore: fix a bunch of cargo warnings & errors
