+++ polars/polars-lazy/polars-plan/src/dsl/string.rs
@@ -1,6 +1,7 @@
 use polars_arrow::array::ValueSize;
 #[cfg(feature = "dtype-struct")]
 use polars_arrow::export::arrow::array::{MutableArray, MutableUtf8Array};
+#[cfg(feature = "dtype-struct")]
 use polars_utils::format_smartstring;
 
 use super::function_expr::StringFunction;
