+++ polars/polars-lazy/polars-plan/src/logical_plan/mod.rs
@@ -269,16 +269,6 @@ impl Default for LogicalPlan {
 }
 
 impl LogicalPlan {
-    #[cfg(test)]
-    pub(crate) fn into_alp(self) -> (Node, Arena<ALogicalPlan>, Arena<AExpr>) {
-        let mut lp_arena = Arena::with_capacity(16);
-        let mut expr_arena = Arena::with_capacity(16);
-        let root = to_alp(self, &mut expr_arena, &mut lp_arena).unwrap();
-        (root, lp_arena, expr_arena)
-    }
-}
-
-impl LogicalPlan {
     pub fn schema(&self) -> PolarsResult<Cow<'_, SchemaRef>> {
         use LogicalPlan::*;
         match self {
@@ -318,122 +308,5 @@ impl LogicalPlan {
     }
     pub fn describe(&self) -> String {
         format!("{self:#?}")
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use polars_core::df;
-    use polars_core::prelude::*;
-
-    use crate::prelude::*;
-    use crate::tests::get_df;
-
-    fn print_plans(lf: &LazyFrame) {
-        println!("LOGICAL PLAN\n\n{}\n", lf.describe_plan());
-        println!(
-            "OPTIMIZED LOGICAL PLAN\n\n{}\n",
-            lf.describe_optimized_plan().unwrap()
-        );
-    }
-
-    #[test]
-    fn test_lazy_arithmetic() {
-        let df = get_df();
-        let lf = df
-            .lazy()
-            .select(&[((col("sepal.width") * lit(100)).alias("super_wide"))])
-            .sort("super_wide", SortOptions::default());
-
-        print_plans(&lf);
-
-        let new = lf.collect().unwrap();
-        println!("{:?}", new);
-        assert_eq!(new.height(), 7);
-        assert_eq!(
-            new.column("super_wide").unwrap().f64().unwrap().get(0),
-            Some(300.0)
-        );
-    }
-
-    #[test]
-    fn test_lazy_logical_plan_filter_and_alias_combined() {
-        let df = get_df();
-        let lf = df
-            .lazy()
-            .filter(col("sepal.width").lt(lit(3.5)))
-            .select(&[col("variety").alias("foo")]);
-
-        print_plans(&lf);
-        let df = lf.collect().unwrap();
-        println!("{:?}", df);
-    }
-
-    #[test]
-    fn test_lazy_logical_plan_schema() {
-        let df = get_df();
-        let lp = df
-            .clone()
-            .lazy()
-            .select(&[col("variety").alias("foo")])
-            .logical_plan;
-
-        assert!(lp.schema().unwrap().get("foo").is_some());
-
-        let lp = df
-            .lazy()
-            .groupby([col("variety")])
-            .agg([col("sepal.width").min()])
-            .logical_plan;
-        assert!(lp.schema().unwrap().get("sepal.width").is_some());
-    }
-
-    #[test]
-    fn test_lazy_logical_plan_join() {
-        let left = df!("days" => &[0, 1, 2, 3, 4],
-        "temp" => [22.1, 19.9, 7., 2., 3.],
-        "rain" => &[0.1, 0.2, 0.3, 0.4, 0.5]
-        )
-        .unwrap();
-
-        let right = df!(
-        "days" => &[1, 2],
-        "rain" => &[0.1, 0.2]
-        )
-        .unwrap();
-
-        // check if optimizations succeeds without selection
-        {
-            let lf = left
-                .clone()
-                .lazy()
-                .left_join(right.clone().lazy(), col("days"), col("days"));
-
-            print_plans(&lf);
-            // implicitly checks logical plan == optimized logical plan
-            let _df = lf.collect().unwrap();
-        }
-
-        // check if optimization succeeds with selection
-        {
-            let lf = left
-                .clone()
-                .lazy()
-                .left_join(right.clone().lazy(), col("days"), col("days"))
-                .select(&[col("temp")]);
-
-            let _df = lf.collect().unwrap();
-        }
-
-        // check if optimization succeeds with selection of a renamed column due to the join
-        {
-            let lf = left
-                .lazy()
-                .left_join(right.lazy(), col("days"), col("days"))
-                .select(&[col("temp"), col("rain_right")]);
-
-            print_plans(&lf);
-            let _df = lf.collect().unwrap();
-        }
     }
 }
